Fix running in a subdirectory of the project root
GITHUB_WORKSPACE is clobbered by GitHub if you try to set it yourself,
so it's useless for allowing a user to override the working directory.
Changing the name of the variable fixes this.

We don't need to use GITHUB_WORKSPACE as a fallback because GitHub
actions already start us off in the project root by default.