feat: add node version control
This will allow the user to control which version of node they
want to use when they run the action.

The current behaviour locks the node version to `latest` which
may not be ideal for certain scenarios.

The setting is optional and defaults to `latest` so it should not
break any exising packages without the option set.